chore: refactor handling of different response errors to separate functions
